Split index route components
